SG-719: Lock the News paragraph to always use `block_1`.
Currently this only affects one node (the homepage), and there is no difference in the field value, which currently references `block_1.`  Since we created another Views display in a previous commit, that would also become a selectable option in the configuration of this Paragrpah. This prevent’s that option from appearing and makes the UI simpler as a result.