sg-1105 - Add new “Content top” field to quicky allow more placement flexibility for “Topic” content.
It is my hope that this is a VERY temporary implementation, as the proper way to solve this is to make all paragaphs available in field_content. However, this was not wanted, as content changed (or migration) would be needed, and this is an urgent need.  See this thread for more details: https://sfdigitalservices.slack.com/archives/CBXM1V2QK/p1591204379001200