sg-1098 - Remove maxlength customizations. They’re not uniformly possible…
The markup is different, even for the same widget, for unknown reasons. For, example, some textarea widgets will include a description inside the .text-format-wrapper, others it will be completely outside of it, while the textarea is in a nested widget. 😭 Ultimately, this template https://git.drupalcode.org/project/drupal/-/blob/9.0.x/core/themes/classy/templates/content-edit/text-format-wrapper.html.twig#L16, would be the one we need to target to style this properly, but it does not have pluggable attributes on the wrapper!!! 😩