Update Drupal Core, Simplify and Update Dev Dependencies... (#712)
* Update core, simplify and update dev dependencies, fix scaffolding…

prep for Composer 2, remove unneessary packages, and add:

- zaporylie/composer-drupal-optimizations: improves performance for Composer 1.  Should be removed for Composer 2.
- drupal/coder and dealerdirect/phpcodesniffer-composer-installer for local code style.

* Add .gitattributes.

Pantheon is free to ignore it, as it is below the cut line, but we should have this for Github.

* Fix robots.txt exclude, which needs mapping.

* Use the real project name.

* Patch gitattributes with support for linguist, which provides….

proper syntax highlighting in Github diffs for non standard php files, like .module, and .install.  Doing this via patch so that any additional changes that come through the scaffold will still come through.