[SG-1724] - Document paragraphs only show the first file on a multi item file field. (#1299)
This is a rebuild to the output of the document paragraph. The preprocess function now builds an array of file information
to print out in a loop on the twig template. Before the twig template was setup to print the first, and only first item of the
files field.

Instructions:
1. Clear cache
2. Review a transaction node with multiple files in a document paragraph.
3. Confirm visually that all documents print out, and thru the inspector, that the documents all list out in an unordered list.