[SG-1738] - Auto select radio buttons for "Request Public Record" on Departments.
Because of the unique setup between the Request public record radio/checkbox and the related phone/email/link fields that go with it, The standard translation setup would not work with tmgmt.

To keep the radios in sync, but still allow the other fields to render and be editable, the TMGMT translation for that field needed to be disabled. Then the default language value for that field needed to be reapplied and locked down on the node translation add/edit pages.

Instructions:
1. config:import sync the field_req_public_records yml file and confirm that the field is in sync (confirm that tmgmt translations is exluded on the field settings page)
2. Clear cache.
3. Complete tmgmt translation steps for a department node on the ticket - https://sfgovdt.jira.com/browse/SG-1738