[SG-1007] - Changed the events view to process the Node ID when filtering the topic and department upcoming/past events view pages.
The previous version couldn't handle special characters like & and ( or ) being in a title, and be able to successfully query the db to find a matching title to the node.
Changing to NID allows the view to contextually filter without issue.

Added a title and target to the "See more events" link generated by the upcoming events blocks. This should help with accessibility for the NID style urls.