[SG-1007] - Better automatic creation of aliases for the not people friendly contextual urls for filtered upcoming and past-events
(Example: /events/topics/25 -> /events/topics/immigrants). An alias is created for language when a topic or department node is visited and a more events link is created. This will generate the aliases only as they are needed as the only manual way to get to the filtered pages are via the "more events" link on those main topic and department nodes.