[SG-1771] - Add keyboard focus indicator to city logo seal and department icons (#1333)
* [SG-1771] - Add keyboard focus indicator to city logo seal and department icons

Setup the logo/icon to have a focus border.
Changed the overall style of focus to a solid dark gray border.
Setup links in dark backgrounds with white text to have a white focus.

Instructions:
1. Clear cache
2. starting at top of page, hit tab until it gets to the logo.
3. Confirm that the logo has a focus border.
4. Continue tabbing till it gets to a dark background element and shows a white colored focus.

* [SG-1771] - Improved setup of white focus color set for dark areas.

* [SG-1771] -  Added white focus for btn-inverse class

* [SG-1771] - Reset to focus items to use the dotted light-gray style until new style is decided on.

Setup to use the @mixin for focus design instead of 3 separate instances of styles. So styles are set now in one place.