[SG-1007] - "See more events" is linking to the wrong list view (#1337)
* [SG-1748] - "See more events" is linking to the wrong list view

Rebuild of all events pages (upcoming and past), and added new page displays for topic and department filtered events (both upcoming and past).

Required reworking of the view templates and preprocess code to account for the new displays.

Reworked the "See more events" link on topic and department pages to automatically go to the filtered event upcoming pages. Filtered past event pages can be clicked to on the upcoming page.

Instructions:
1. clear cache.
2. make sure config is synced for the "Events" view.
3. Visit /events and confirm that upcoming and past events still display as they should.
4. Visit a topic page and a department page, that have at least 4 upcoming events (create some if needed) and click on the "See more events" link
5. Should take you to a url that follows this pattern - /events/topic/(title-of-topic) or /events/department/(title-of-department)
6. Can manually go to topic pages like "/events/topic/immigrants", "/events/topic/taxes", or department versions like "/events/department/treasure-island-development-authority"

Remember, event content may need to be created to be able to see upcoming events because most topics and departments do not have many/any upcoming events.

Notes: It is possible that if a topic or department title contains some abnormal characters (like & or $), it might not process cleanly for the contextual filtering.
Some known fixes are in place, but stuff is possible, however unlikely it may be.

* [SG-1007] - Changed the events view to process the Node ID when filtering the topic and department upcoming/past events view pages.

The previous version couldn't handle special characters like & and ( or ) being in a title, and be able to successfully query the db to find a matching title to the node.
Changing to NID allows the view to contextually filter without issue.

Added a title and target to the "See more events" link generated by the upcoming events blocks. This should help with accessibility for the NID style urls.

* [SG-1007] - Fix php error occurring if an empty node gets processed in the events filtering

* [SG-1007] - Better automatic creation of aliases for the not people friendly contextual urls for filtered upcoming and past-events

(Example: /events/topics/25 -> /events/topics/immigrants). An alias is created for language when a topic or department node is visited and a more events link is created. This will generate the aliases only as they are needed as the only manual way to get to the filtered pages are via the "more events" link on those main topic and department nodes.

* [SG-1007] - Moved Alias generating code out of theme preprocess layer and into utilities module. Added a check to see if the system alias needs to be updated due to changes in the desired alias path.

* [SG-1007] -- Trying to fix php error

* SG-1007 -- Fix issue with redirect checks on nodes.

For departments with events, this was causing issues with the events display when trying to view upcoming and past events for that department.
The redirect check was not receiving a full node object to process on that display of department events.

* SG-1007 -- Fix to see more department/topics event link issue not registering node. Better node object checking.

* Removed target attribute

Removed target attribute from link.