[SG-1814] -- Extra spacing in bios trimmed "teaser" content caused by extra p and br tags
#SG-1814

The text format sf_restricted_html was being applied to the trimmed teaser of the bios, and is not setup to handle br tags, which caused
extra paragraph tags to be generated for the output. these were causing the extra spacing on the trimmed display. The expanded content was
rendered using full_html so it was not generating the break tags or extra paragraph tags. That is why the two content sets appeared different
when it was collapsed and expanded.

Instructions
1. Make sure caches are cleared
2. Review a bio with expanding bio content.
3. Confirm that the content no longer has extra spacing.