[SG-1894] - Truncation in a description when a campaign is used in Related.
SG-1894

This setup changes the way it grabs the summary from an About section. The character length was increased to potentially grab a larger paragraph, and then a snippet was added to try to pull out the first <p></p> content in the about text. This will need to be checked, and maybe some About content will still need to be edited, in order for this to best work for all cases.

Instructions:
- Config import
- Clear cache
- Visit a information page, or any type that allows for campaigns to be added to the Resource sidebar
- Confirm that the text is no longer cutoff

Note:
- Edit any campaign about sections to make sure the first about paragraph fits in the character limits.