[SG-1915] -- Do not show "draft" agency pages on the Division list. (#1460)
SG-1915

Agency nodes that are in draft state, with no existing published version accessible to the public, are now set to not display in the about section (under the divisions/subcommittee section)
It does this by checking the current moderation state of the division, and if it is "draft" then it does not allow it in the list. It applies this check for all users and roles, so not even admins will see the draft link.

Instructions:
Cache clear
Create and new agency and only assign to draft state (DO NOT publish any state for this test agency)
Add this draft agency to the about section of an agency
Review this page as any user (admin, logged in, anonymous, private browser). The link to the test draft agency should not display

NOTES: there are comments in the code for if this ever needs to change to block both archived and completely draft states.