[SG-1947] -- Fixed Data story "Part of..." tagging in the header.
SG-1947

This fixes the Part of tagging setup that finds resources that make reference to a data story, and then prints that reference on the data story.
The issue was that past versions of the content made it looked like it still referenced content. This fix adds a double check to make sure the
latest version of the resource does indeed still reference back to the data story we are on.

Instructions
- clear cache
- open a data story with multiple "Part of..." tags and open the resources in those tags in separate tabs
- edit one of the resources to not reference back to the data story and save.
- review the data story and confirm that the resource is now removed from the "Part of..." element.