[SG-1369] -- Remove duplicate publication status bars on event full pages (#1580)
Event full pages render the content moderation bar twice (header and just above body field). The content moderation field needed to be added to the list of items to not render in the content variable on the template. This item is specifically called out to be rendered in the header, so it does not need to be included in the generic content render.

Instructions
- clear cache
- visit any draft/unpublished event node as an editor
- confirm that the moderation bar only renders once